Sex maximum width for investigatory report image on petaluma poultry case page.
Before this change, the image would resize to the screen width, and the text in the
image could become too large compared to other text on the page. An alternative to
this change would have been to use an image with no text and overlay it with text
in HTML that would remain proportional to other text.