Add petition to Petaluma Poultry page
Zoe said she doesn't want to refer people to perdueabuse.com if there is
no petition on there for people to sign.

Add support in header component for alternative targets so the page
does not refresh when clicking a button that links to something further
down on the same page.